Added heatmap overlay. Calibrated eye-tracker values
**Added**
Heatmap overlay/background background color. We can now see the background in the heatmap as the vscode window. The intensity color of the heatmap is now from transparent to red to white (white is most intense)

**Calibrated eye-tracker values**
Eye-tracker values now get better accuracy for Hugo's computer. Now we have both for Eric's and Hugo's screens.

Hugo Collaboration with Eric

Co-Authored-By: FScirE <72918823+FScirE@users.noreply.github.com>